# api/swagger/openapi.yml
openapi: "3.0.3"
info:
  title: GophKeeper API
  description: Password manager REST API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a new user
      operationId: RegisterUser
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/registerRequest'
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        "400":
          $ref: '#/components/responses/badRequest'
        "409":
          $ref: '#/components/responses/conflict'
  /api/v1/auth/login:
    post:
      summary: Login and get tokens
      operationId: LoginUser
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginRequest'
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tokenResponse'
        "400":
          $ref: '#/components/responses/badRequest'
        "401":
          $ref: '#/components/responses/unauthorized'
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      operationId: RefreshToken
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/refreshRequest'
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/tokenResponse'
        "400":
          $ref: '#/components/responses/badRequest'
        "401":
          $ref: '#/components/responses/unauthorized'
  /api/v1/data:
    post:
      summary: Create a new data entry
      operationId: CreateData
      tags:
        - data
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createDataRequest'
      responses:
        "201":
          description: Data entry created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/createDataResponse'
        "400":
          $ref: '#/components/responses/badRequest'
  /api/v1/data/list:
    get:
      summary: List data entries with filtering and pagination
      operationId: ListData
      tags:
        - data
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: types
          in: query
          description: Filter by entry types (comma-separated)
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/dataType'
            style: form
            explode: false
        - name: query
          in: query
          description: Search query (searches by metadata)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of data entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listDataResponse'
        "400":
          $ref: '#/components/responses/badRequest'
  /api/v1/data/{id}:
    get:
      summary: Get a data entry by ID
      operationId: GetData
      tags:
        - data
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Data entry ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Data entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dataInfo'
        "404":
          $ref: '#/components/responses/notFound'

    put:
      summary: Update a data entry
      operationId: UpdateData
      tags:
        - data
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Data entry ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateDataRequest'
      responses:
        "200":
          description: Data entry updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/updateDataResponse'
        "400":
          $ref: '#/components/responses/badRequest'
        "404":
          $ref: '#/components/responses/notFound'

    delete:
      summary: Delete a data entry
      operationId: DeleteData
      tags:
        - data
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Data entry ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Data entry deleted
        "404":
          $ref: '#/components/responses/notFound'
  /api/v1/file/upload/init:
    post:
      summary: Initialize a new file upload
      operationId: InitUpload
      tags:
        - file
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/fileInitRequest'
      responses:
        "201":
          description: File upload initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fileInitResponse'
        "400":
          $ref: '#/components/responses/badRequest'
  /api/v1/file/upload/chunk:
    post:
      summary: Upload a file chunk
      operationId: UploadChunk
      tags:
        - file
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file_id
                - chunk_index
                - data
              properties:
                file_id:
                  type: string
                  format: uuid
                chunk_index:
                  type: integer
                  minimum: 0
                data:
                  type: string
                  format: binary
      responses:
        "200":
          description: Chunk uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fileChunkResponse'
        "400":
          $ref: '#/components/responses/badRequest'
        "404":
          $ref: '#/components/responses/notFound'
  /api/v1/file/upload/complete:
    post:
      summary: Complete file upload
      operationId: CompleteUpload
      tags:
        - file
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/fileCompleteRequest'
      responses:
        "200":
          description: File upload completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/fileInfo'
        "400":
          $ref: '#/components/responses/badRequest'
        "404":
          $ref: '#/components/responses/notFound'
  /api/v1/file/{id}/download:
    get:
      summary: Download a file
      operationId: DownloadFile
      tags:
        - file
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: File ID
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          $ref: '#/components/responses/notFound'
  /api/v1/file/{id}:
    delete:
      summary: Delete a file
      operationId: DeleteFile
      tags:
        - file
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: File ID
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: File deleted
        "404":
          $ref: '#/components/responses/notFound'
  /api/v1/file/list:
    get:
      summary: List user's files with pagination and search
      operationId: ListFiles
      tags:
        - file
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: page_size
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: query
          in: query
          description: Search query (searches by file name)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/listFilesResponse'
        "400":
          $ref: '#/components/responses/badRequest'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    badRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    notFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
...
    registerRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 4
          maxLength: 128
    loginRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
        password:
          type: string
    refreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    tokenResponse:
      type: object
      required:
        - token_type
        - access_token
        - refresh_token
        - expires_in
        - refresh_token_expires_in
      properties:
        token_type:
          type: string
          example: "bearer"
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Access token expiration time in seconds
        refresh_token_expires_in:
          type: integer
          description: Refresh token expiration time in seconds
    user:
      type: object
      required:
        - id
        - login
        - created_at
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
        created_at:
          type: string
          format: date-time
    createDataRequest:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/data'
    createDataResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/dataInfo'
    updateDataRequest:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/data'
    updateDataResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/dataInfo'
    dataType:
      type: string
      enum:
        - LOGIN_PASSWORD
        - BANK_CARD
        - TEXT
        - FILE
      description: Type of data entry
    dataInfo:
      type: object
      oneOf:
        - $ref: '#/components/schemas/loginPasswordDataInfo'
        - $ref: '#/components/schemas/bankCardDataInfo'
        - $ref: '#/components/schemas/textDataInfo'
        - $ref: '#/components/schemas/fileDataInfo'
      discriminator:
        propertyName: type
        mapping:
          LOGIN_PASSWORD: '#/components/schemas/loginPasswordDataInfo'
          BANK_CARD: '#/components/schemas/bankCardDataInfo'
          TEXT: '#/components/schemas/textDataInfo'
          FILE: '#/components/schemas/fileDataInfo'
...
    loginPasswordDataInfo:
      type: object
      allOf:
        - $ref: '#/components/schemas/baseDataInfo'
        - $ref: '#/components/schemas/loginPasswordData'
...
    listDataResponse:
      type: object
      required:
        - items
        - total
        - page
        - page_size
        - total_pages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/dataInfo'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...
    fileInitRequest:
      type: object
      required:
        - name
        - mime_type
        - size
        - chunks_count
      properties:
        name:
          type: string
          maxLength: 1024
        mime_type:
          type: string
          maxLength: 256
        size:
          type: integer
          format: int64
          minimum: 0
        chunks_count:
          type: integer
          minimum: 1
    fileInitResponse:
      type: object
      required:
        - file_id
        - chunks_count
      properties:
        file_id:
          type: string
          format: uuid
        chunks_count:
          type: integer
    fileChunkResponse:
      type: object
      required:
        - file_id
        - chunk_index
        - uploaded
      properties:
        file_id:
          type: string
          format: uuid
        chunk_index:
          type: integer
        uploaded:
          type: boolean
    fileCompleteRequest:
      type: object
      required:
        - file_id
      properties:
        file_id:
          type: string
          format: uuid
    fileInfo:
      type: object
      required:
        - id
        - name
        - mime_type
        - size
        - chunks_count
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        mime_type:
          type: string
        size:
          type: integer
          format: int64
        chunks_count:
          type: integer
        status:
          $ref: '#/components/schemas/fileStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    listFilesResponse:
      type: object
      required:
        - items
        - total
        - page
        - page_size
        - total_pages
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/fileInfo'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer